+++ src/Tests/Storage.fs
@@ -3,6 +3,7 @@ module Storage
 open Expecto
 open Farmer
 open Farmer.Builders
+open Farmer.Network
 open Farmer.Storage
 open Microsoft.Azure.Management.Storage
 open Microsoft.Azure.Management.Storage.Models
@@ -209,5 +210,32 @@ let tests = testList "Storage Tests" [
                     default_blob_access_tier Cool
                 } |> ignore)
             "Can't set default tier for  Block Blobs"
+    }
+    test "Restrict by IP" {
+        let storage = storageAccount {
+            name "onlymyhouse24125"
+            restrict_to_ip "8.8.8.8"
+            restrict_to_prefix "8.8.8.0/24"
+        }
+        let generated = arm { add_resource storage; } |> getStorageResource
+        Expect.hasLength generated.NetworkRuleSet.IpRules 2 "Wrong number of IP rules"
+        Expect.containsAll (generated.NetworkRuleSet.IpRules |> Seq.map (fun rule -> rule.IPAddressOrRange)) [ "8.8.8.8"; "8.8.8.0/24" ] "Missing IP rules"
+    }
+    test "Restrict to vnet" {
+        let vnetName = "my-vnet"
+        let servicesSubnet = "services"
+        let containerSubnet = "containers"
+        let storage = storageAccount {
+            name "onlymynet"
+            restrict_to_subnet vnetName servicesSubnet
+            restrict_to_subnet vnetName containerSubnet
+        }
+        let generatedStorage = arm { add_resource storage; } |> getStorageResource
+        Expect.hasLength generatedStorage.NetworkRuleSet.VirtualNetworkRules 2 "Wrong number of vnet rules"
+        let allowedSubnets = [
+            (Arm.Network.subnets.resourceId (ResourceName vnetName, ResourceName servicesSubnet)).ArmExpression.Eval()
+            (Arm.Network.subnets.resourceId (ResourceName vnetName, ResourceName containerSubnet)).ArmExpression.Eval()
+        ]
+        Expect.containsAll allowedSubnets (generatedStorage.NetworkRuleSet.VirtualNetworkRules |> Seq.map (fun rule -> rule.VirtualNetworkResourceId)) "Missing subnet rules"
     }
 ]